+++ src/parray.c
@@ -36,7 +36,7 @@
 #include "internal/arrays.h"
 #include "internal/collections.h"
 // ------------------------------
-#include <stdlib.h>
+#include <sigma.core/allocator.h>
 #include <string.h>
 
 //  declare the PointerArray struct: the collection with attitude
@@ -90,9 +90,9 @@ static void parray_init(parray *arr, usize capacity) {
         // what to do about an array that's already initialized?
         // for now, we just reallocate the bucket
         if ((*arr)->bucket) {
-            coll_free((*arr)->bucket);
+            Allocator.dispose((*arr)->bucket);
         }
-        (*arr)->bucket = coll_alloc(sizeof(addr) * capacity);
+        (*arr)->bucket = Allocator.alloc(sizeof(addr) * capacity);
         if (!(*arr)->bucket) {
             // allocation ERRed, handle error as needed
             return;
@@ -211,5 +211,4 @@ const sc_parray_i PArray = {
     .as_collection = parray_as_collection,
     .to_collection = parray_to_collection,
     .as_slotarray = parray_as_slotarray,
-    .alloc_use = coll_set_alloc_use,
 };